# Designs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "IteraWebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Blogs/StoreEntityAsync": {
      "post": {
        "tags": [
          "Blogs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Blog"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Blog"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Blog"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Blogs/GetEntityByIdAsync": {
      "post": {
        "tags": [
          "Blogs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Blogs/DeleteEntityAsync": {
      "post": {
        "tags": [
          "Blogs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanParticipants/AddParticipantAsync": {
      "post": {
        "tags": [
          "LeanParticipants"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeanParticipant"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LeanParticipant"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LeanParticipant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanParticipants/GetEntityByIdAsync": {
      "post": {
        "tags": [
          "LeanParticipants"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanParticipants/DeleteEntityAsync": {
      "post": {
        "tags": [
          "LeanParticipants"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/StoreEntityAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeanSession"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LeanSession"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LeanSession"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/GetEntityByIdAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetEntityByIdQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/GetLeanSessionAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLeanSessionQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLeanSessionQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetLeanSessionQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/GetLeanSessionsAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLeanSessionsQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetLeanSessionsQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetLeanSessionsQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/CloseSessionAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseSessionCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseSessionCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloseSessionCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/StoreNoteAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreLeanSessionNoteCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreLeanSessionNoteCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StoreLeanSessionNoteCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/LeanSessions/DeleteEntityAsync": {
      "post": {
        "tags": [
          "LeanSessions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteEntityCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Blog": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedBy": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "nullable": true
          },
          "featuredImageUrl": {
            "type": "string",
            "nullable": true
          },
          "abstract": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CloseSessionCommand": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeleteEntityCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "entityId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetEntityByIdQuery": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "entityId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetLeanSessionQuery": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetLeanSessionsQuery": {
        "type": "object",
        "properties": {
          "searchTerm": {
            "type": "string",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "facilitatorUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LeanParticipant": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedBy": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "leanSessionId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/ParticipantRole"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "leftAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "LeanSession": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedBy": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "scheduledStartTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualStartTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualEndTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "facilitatorUserId": {
            "type": "string",
            "nullable": true
          },
          "defaultTopicDuration": {
            "type": "integer",
            "format": "int32"
          },
          "isPublic": {
            "type": "boolean"
          },
          "inviteCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NoteType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ParticipantRole": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SessionStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "StoreLeanSessionNoteCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "leanSessionId": {
            "type": "string",
            "nullable": true
          },
          "leanTopicId": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "noteType": {
            "$ref": "#/components/schemas/NoteType"
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}